atlas-persistence: move dedup and file checking to inside stream operator for better sync, also avoid a potential leak at activeFiles
