+++ atlas-persistence/src/main/scala/com/netflix/atlas/persistence/S3CopyService.scala
@@ -16,7 +16,6 @@
 package com.netflix.atlas.persistence
 
 import java.io.File
-import java.util.concurrent.ConcurrentHashMap
 
 import akka.NotUsed
 import akka.actor.ActorSystem
@@ -61,48 +60,15 @@ class S3CopyService @Inject()(
     "`max-inactive-duration` MUST be longer than `max-duration`, otherwise file may be renamed before normal write competes"
   )
 
-  // Tracking files which is being processed to avoid duplication
-  private val activeFiles = new ConcurrentHashMap[String, KillSwitch]
-
-  // TODO handle .tmp
   override def startImpl(): Unit = {
     logger.info("Starting service")
     killSwitch = Source
       .tick(1.second, 5.seconds, NotUsed)
       .viaMat(KillSwitches.single)(Keep.right)
       .flatMapMerge(Int.MaxValue, _ => Source(FileUtil.listFiles(new File(dataDir))))
-      .filter(shouldProcess)
       // S3CopySink handles flow restart for each file so no need to use RestartSink here
-      .toMat(new S3CopySink(bucket, region, prefix, activeFiles, registry, system))(Keep.left)
+      .toMat(new S3CopySink(bucket, region, prefix, maxInactiveMs, registry, system))(Keep.left)
       .run()
-  }
-
-  def shouldProcess(f: File): Boolean = {
-    if (activeFiles.containsKey(f.getName)) {
-      logger.debug(s"Should NOT process: being processed - $f")
-      false
-    } else if (FileUtil.isTmpFile(f)) {
-      if (isInactive(f)) {
-        logger.warn(s"Should process: temp file but inactive - $f")
-        true
-      } else {
-        logger.debug(s"Should NOT process: temp file - $f")
-        false
-      }
-    } else {
-      logger.debug(s"Should process: regular file - $f")
-      true
-    }
-  }
-
-  private def isInactive(f: File): Boolean = {
-    try {
-      System.currentTimeMillis > f.lastModified() + maxInactiveMs
-    } catch {
-      case e: Exception =>
-        logger.error(s"Error get lastModified for file $f", e)
-        false
-    }
   }
 
   override def stopImpl(): Unit = {
