+++ atlas-persistence/src/main/scala/com/netflix/atlas/persistence/S3CopySink.scala
@@ -41,13 +41,15 @@ import software.amazon.awssdk.services.s3.S3AsyncClient
 import software.amazon.awssdk.services.s3.model.PutObjectRequest
 import software.amazon.awssdk.services.s3.model.PutObjectResponse
 
+import scala.concurrent.duration._
 import scala.jdk.CollectionConverters._
+import scala.jdk.FutureConverters._
 
 class S3CopySink(
   val bucket: String,
   val region: String,
   val prefix: String,
-  val activeFiles: ConcurrentHashMap[String, KillSwitch],
+  val maxInactiveMs: Long,
   val registry: Registry,
   implicit val system: ActorSystem
 ) extends GraphStage[SinkShape[File]]
@@ -63,9 +65,9 @@ class S3CopySink(
     new GraphStageLogic(shape) with InHandler {
 
       @volatile private var s3Client: S3AsyncClient = _
-      // This map tracks ongoing file copy flows. It should be thread safe since "remove" happens
-      // in a different flow.
       private val numActiveFiles = registry.distributionSummary("persistence.s3.numActiveFiles")
+      // Tracking files which is being processed to avoid duplication
+      private val activeFiles = new ConcurrentHashMap[String, Option[KillSwitch]]
 
       override def preStart(): Unit = {
         initS3Client()
@@ -73,18 +75,21 @@ class S3CopySink(
       }
 
       override def onPush(): Unit = {
-        process(grab(in))
+        val file = grab(in)
+        if (shouldProcess(file)) {
+          process(file)
+        }
         pull(in)
       }
 
       override def onUpstreamFinish(): Unit = {
         super.completeStage()
-        activeFiles.values.asScala.foreach(_.shutdown())
+        activeFiles.values.asScala.foreach(option => option.foreach(_.shutdown()))
       }
 
       override def onUpstreamFailure(ex: Throwable): Unit = {
         super.failStage(ex)
-        activeFiles.values.asScala.foreach(_.shutdown())
+        activeFiles.values.asScala.foreach(option => option.foreach(_.shutdown()))
       }
 
       setHandler(in, this)
@@ -96,15 +101,43 @@ class S3CopySink(
           .build()
       }
 
+      def shouldProcess(f: File): Boolean = {
+        if (activeFiles.containsKey(f.getName)) {
+          logger.debug(s"Should NOT process: being processed - $f")
+          false
+        } else if (FileUtil.isTmpFile(f)) {
+          if (isInactive(f)) {
+            logger.warn(s"Should process: temp file but inactive - $f")
+            true
+          } else {
+            logger.debug(s"Should NOT process: temp file - $f")
+            false
+          }
+        } else {
+          logger.debug(s"Should process: regular file - $f")
+          true
+        }
+      }
+
+      private def isInactive(f: File): Boolean = {
+        try {
+          System.currentTimeMillis > f.lastModified() + maxInactiveMs
+        } catch {
+          case e: Exception =>
+            logger.error(s"Error get lastModified for file $f", e)
+            false
+        }
+      }
+
       // Start a new stream to copy each file
       private def process(file: File): Unit = {
-        import scala.concurrent.duration._
-        import scala.jdk.FutureConverters._
+        // Adding "None" just to indicate if cleanupFile has been called
+        activeFiles.put(file.getName, None)
 
         val killSwitch = RestartSource
           .onFailuresWithBackoff(
             minBackoff = 1.seconds,
-            maxBackoff = 5.seconds,
+            maxBackoff = 1.seconds,
             randomFactor = 0,
             maxRestarts = -1
           ) { () =>
@@ -114,7 +147,10 @@ class S3CopySink(
           .toMat(Sink.foreach(_ => cleanupFile(file)))(Keep.left)
           .run
 
-        activeFiles.put(file.getName, killSwitch)
+        // Note: computeIfPresent makes sure the file is only added to map if it's present (before
+        //   cleanFile removes it) to avoid a leak
+        activeFiles.computeIfPresent(file.getName, (k, v) => Option(killSwitch))
+
         numActiveFiles.record(activeFiles.size)
       }
